+++ Helper/Data.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright ©  All rights reserved.
+ * Copyright © Happy Horizon Utrecht Development & Technology B.V. All rights reserved.
  * See COPYING.txt for license details.
  */
 declare(strict_types=1);
@@ -13,23 +13,22 @@ use Magento\Framework\App\Helper\Context;
 use Magento\Framework\Exception\FileSystemException;
 use Magento\Framework\Filesystem\Driver\File;
 use Magento\Framework\Filesystem\DirectoryList;
-use Magento\Backend\Controller\Adminhtml\Cache\MassRefresh;
 use Magento\Framework\Serialize\Serializer\Json;
 
 class Data extends AbstractHelper
 {
     public const GQL_FILE_NAME = 'gql.php';
+
     /**
      * @param Context $context
      * @param DirectoryList $dir
      * @param File $file
-     * @param MassRefresh $massRefresh
+     * @param Json $json
      */
     public function __construct(
-        protected Context $context,
+        Context $context,
         protected DirectoryList $dir,
         protected File $file,
-        protected MassRefresh $massRefresh,
         protected Json $json
     ) {
         parent::__construct($context);
@@ -41,17 +40,17 @@ class Data extends AbstractHelper
     public function getGqlPath(): string
     {
         try {
-            return $this->dir->getPath(DirectoryListApp::CONFIG). DIRECTORY_SEPARATOR .self::GQL_FILE_NAME;
+            return $this->dir->getPath(DirectoryListApp::CONFIG) . DIRECTORY_SEPARATOR . self::GQL_FILE_NAME;
         } catch (FileSystemException $e) {
             return '';
         }
     }
 
     /**
-     * @param $path
+     * @param string $path
      * @return bool
      */
-    public function checkIfFileExists($path): bool
+    public function checkIfFileExists(string $path): bool
     {
         try {
             return $this->file->isExists($path);
@@ -61,10 +60,10 @@ class Data extends AbstractHelper
     }
 
     /**
-     * @param $path
+     * @param string $path
      * @return bool
      */
-    public function removeFile($path): bool
+    public function removeFile(string $path): bool
     {
         try {
             if ($this->checkIfFileExists($path)) {
@@ -78,11 +77,11 @@ class Data extends AbstractHelper
     }
 
     /**
-     * @param $path
-     * @param $content
+     * @param string $path
+     * @param string $content
      * @return bool
      */
-    public function saveFileContent($path, $content): bool
+    public function saveFileContent(string $path, string $content): bool
     {
         try {
             if ($this->checkIfFileExists($path)) {
@@ -96,34 +95,35 @@ class Data extends AbstractHelper
     }
 
     /**
-     * @param $path
-     * @return string|void
+     * @param string $path
+     * @return string
      */
-    public function getFileContent($path)
+    public function getFileContent(string $path): string
     {
         try {
             if ($this->checkIfFileExists($path)) {
-                $this->file->fileGetContents($path);
+                return $this->file->fileGetContents($path);
             }
         } catch (FileSystemException $e) {
-            return "";
+            return '';
         }
+        return '';
     }
 
     /**
-     * @param $data
-     * @return bool|string
+     * @param mixed $data
+     * @return string
      */
-    public function encodeData($data): bool|string
+    public function encodeData(mixed $data): string
     {
         return $this->json->serialize($data);
     }
 
     /**
-     * @param $data
-     * @return array|bool|float|int|mixed|string|null
+     * @param string $data
+     * @return mixed
      */
-    public function decodeData($data): mixed
+    public function decodeData(string $data): mixed
     {
         return $this->json->unserialize($data);
     }